moved generation strategies from conftest.py to test_property_based.py. moved file test_property_based.py to unit folder.
